Fixed updateRecipe error

+++ backend/controller/recipe-controller.js
@@ -1,4 +1,4 @@
-const fs = require('fs')
+
 const {validationResult} = require('express-validator')
 const mongoose = require('mongoose')
 const uuid = require('uuid/v4') ;
@@ -130,55 +130,98 @@ const addRecipe = async (req, res, next) => {
         return next(new httpError('Invalid input passed.', 422))
     }
 
-    const type = mime.lookup(req.file.originalname);
-    const gc = new Storage({
-        keyFilename: path.join(__dirname, "../recipe-app-273623-1d4d668a2ea8.json"),
-        projectId: process.env.GOOGLE_PROJECT_ID
-      });
-      
-    const bucket = gc.bucket(process.env.BUCKET_NAME) 
-      
-	
-	const blob = bucket.file(`${uuid()}.${mime.extensions[type][0]}`);
-
-	const stream = blob.createWriteStream({
-		resumable: true,
-		contentType: type,
-		predefinedAcl: 'publicRead',
-	});
-
-	
-    const {title, description, difficulty, cookingTime, preparationTime, category, ingredients, directions, servings, creator } = req.body
-   
-    const newRecipe = new Recipe ({
-        title,
-        description, 
-        imageURL : `https://storage.googleapis.com/${bucket.name}/${blob.name}`,
-        difficulty,
-        cookingTime,
-        preparationTime, 
-        ratings : {
-            averageRating : 0 , 
-            ratings : []
-        },
-        likes : {
-          likesNumber: 0,
-          likes: []
-        },
-        category,
-        ingredients,
-        directions, 
-        servings, 
-        creator
-    })
     let user;
     try {
-        user = await User.findById(creator)
+        user = await User.findById(req.userData.userId)
     } catch(err) {
-        return next(new httpError('finding a user failed'), 500)
+        return next(new httpError('finding a user failed'), 401)
     }
     if (!user) {
-        return next(new httpError('could not find the user for provided id'), 404)
+        return next(new httpError('could not find the user for provided id'), 401)
+    }
+
+    const {
+        title,
+        description,
+        difficulty,
+        cookingTime,
+        preparationTime,
+        category,
+        ingredients,
+        directions,
+        servings
+    } = req.body
+
+    let newRecipe;
+
+    // If image was uploaded
+    if(req.file) {
+        const type = mime.lookup(req.file.originalname);
+        const gc = new Storage({
+            keyFilename: path.join(__dirname, "../recipe-app-273623-1d4d668a2ea8.json"),
+            projectId: process.env.GOOGLE_PROJECT_ID
+        });
+      
+        const bucket = gc.bucket(process.env.BUCKET_NAME) 
+        const blob = bucket.file(`${uuid()}.${mime.extensions[type][0]}`);
+        const stream = blob.createWriteStream({
+            resumable: true,
+            contentType: type,
+            predefinedAcl: 'publicRead',
+        });
+
+        stream.on('error', err => {
+            next(err);
+        });
+
+        stream.on('finish', () => {
+            console.log("Finish stream.")
+        });
+    
+        stream.end(req.file.buffer);
+
+        newRecipe = new Recipe({
+            title,
+            description, 
+            imageURL : `https://storage.googleapis.com/${bucket.name}/${blob.name}`,
+            difficulty,
+            cookingTime,
+            preparationTime, 
+            ratings : {
+                averageRating : 0 , 
+                ratings : []
+            },
+            likes : {
+            likesNumber: 0,
+            likes: []
+            },
+            category,
+            ingredients: JSON.parse(ingredients),
+            directions: JSON.parse(directions), 
+            servings,
+            creator: user.id
+        })
+    } else {
+        newRecipe = new Recipe({
+            title,
+            description,
+            difficulty,
+            cookingTime,
+            preparationTime, 
+            ratings : {
+                averageRating : 0 , 
+                ratings : []
+            },
+            likes : {
+            likesNumber: 0,
+            likes: []
+            },
+            category,
+            ingredients: JSON.parse(ingredients),
+            directions: JSON.parse(directions),  
+            servings, 
+            creator: user.id
+        })
     }
 
     try {
@@ -193,23 +236,9 @@ const addRecipe = async (req, res, next) => {
         return next(new httpError('Adding a recipe failed'), 500)
     }
 
-    stream.on('error', err => {
-		next(err);
-	});
-
-	stream.on('finish', () => {
-		res.status(200).json({
-			data: {
-                url: `https://storage.googleapis.com/${bucket.name}/${blob.name}`,
-                recipe : newRecipe
-			},
-        });
-        
-    });
-
-    stream.end(req.file.buffer);
-
+    res.status(200).json({ recipe: newRecipe });
 }
+
 
 const rateRecipe = async (req, res, next) => {
     const {userId, recipeId, rate} = req.body
@@ -248,18 +277,19 @@ const rateRecipe = async (req, res, next) => {
         else {
             recipe.ratings.ratings.some(function (recipe) { if (recipe.user == userId) { recipe.rating = rate} })
         }
+        // Save new average rating
+        let total = 0;
+        recipe.ratings.ratings.map((recipe) =>{
+            total = total + recipe.rating
+        } )
+        recipe.ratings.averageRating = total / recipe.ratings.ratings.length
         await recipe.save({ session : sess})
         await sess.commitTransaction()
     } catch (err) {
         console.log(err)
         return next(new httpError('Rating a recipe failed'), 500)
     }
-    let total = 0;
-   recipe.ratings.ratings.map((recipe) =>{
-        total = total + recipe.rating
-    } )
-    recipe.ratings.averageRating = total / recipe.ratings.ratings.length
-    console.log(recipe.ratings.averageRating)
+ 
     res.json({ recipe : recipe.toObject({ getters : true }) })
 }
 
@@ -318,7 +348,15 @@ const updateRecipe = async (req,res, next) => {
     }
 
     const recipeId = req.params.recipeId
-    const {title, ingredients, directions} = req.body
+    const {title,
+        description,
+        difficulty,
+        cookingTime,
+        preparationTime,
+        category,
+        ingredients,
+        directions,
+        servings} = req.body
 
     let recipe
     try {
@@ -334,9 +372,15 @@ const updateRecipe = async (req,res, next) => {
     }
 
     recipe.title = title
+    recipe.description = description
+    recipe.difficulty = difficulty
+    recipe.cookingTime = cookingTime
+    recipe.preparationTime = preparationTime
+    recipe.category = category
     recipe.ingredients = ingredients
     recipe.directions = directions
-
+    recipe.servings = servings
+    
     try {
         await recipe.save()
     } catch (err) {
@@ -365,13 +409,7 @@ const deleteRecipe = async (req, res, next) => {
     }
     
     const imagePath = recipe.imageURL
-    const gc = new Storage({
-        keyFilename: path.join(__dirname, "../recipe-app-273623-1d4d668a2ea8.json"),
-        projectId: process.env.GOOGLE_PROJECT_ID
-      });
-    const parts = imagePath.split('/')
-    const filename = parts[(parts.length)-1]
-
+    
     try {
         const sess = await mongoose.startSession();
         sess.startTransaction();
@@ -385,14 +423,23 @@ const deleteRecipe = async (req, res, next) => {
     }
 
     //delete the file from the bucket in google cloud
-    try {
-        await gc.bucket(process.env.BUCKET_NAME).file(filename).delete()
-    } catch (err) {
-        console.log(err)
-        return next(new httpError('Deleting the recipe file from cloud failed'), 500)
+    if(imagePath && imagePath !== "") {
+        const gc = new Storage({
+            keyFilename: path.join(__dirname, "../recipe-app-273623-1d4d668a2ea8.json"),
+            projectId: process.env.GOOGLE_PROJECT_ID
+        });
+        const parts = imagePath.split('/')
+        const filename = parts[(parts.length)-1]
+
+        // Delete file from bucket in google cloud
+        try {
+            await gc.bucket(process.env.BUCKET_NAME).file(filename).delete()
+        } catch (err) {
+            console.log(err)
+            return next(new httpError('Deleting the recipe file from cloud failed'), 500)
+        }
     }
     
-
     res.json({ message : "Deleted recipe"})
 }
